Removed LOWER function from the query
I compiled the application and started the app as per the given instruction with the provided database queries. If I use the LOWER function, it returns null as for the queries as the insert statements keep the data in mixed cases. Either the data should have been updated there or the code here should have been updated. I preferred updating the code as the user may enter the ruleName and other data in the database using any case and the LOWER function will always return null in that case as it was in mine. 
Please review and update the change. Thanks